+++ src/app/addteam/_domain/components/CreateTeamCard.module.css
@@ -1,3 +1,19 @@
+.title {
+  margin: 0;
+  color: var(--color-text-primary);
+  font-size: 24px;
+  font-weight: 700;
+  line-height: 28px;
+}
+
+.label {
+  margin: 0;
+  color: var(--color-text-secondary);
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 19px;
+}
+
 .card {
   position: relative;
   width: var(--create-card-width-desktop);
@@ -26,6 +42,17 @@
 }
 
 @media (max-width: 767px) {
+  .title {
+    font-size: 20px;
+    line-height: 24px;
+  }
+
+  .label {
+    color: var(--color-text-primary);
+    font-size: 14px;
+    line-height: 17px;
+  }
+
   .card {
     width: var(--create-card-width-mobile);
     min-height: 464px;
