+++ src/app/addteam/_domain/components/NoTeamState.tsx
@@ -3,10 +3,10 @@
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { BaseButton } from '@/components/Button/base';
-import noTeamImage from '../svg/noTeamImg.svg';
+import noTeamImage from '../../svg/noTeamImg.svg';
 import noTeamStateStyles from './NoTeamState.module.css';
 import clsx from 'clsx';
-import commonStyles from '../_styles/common.module.css';
+import commonStyles from '../styles/common.module.css';
 
 export default function NoTeamState() {
   const router = useRouter();
