+++ src/app/addteam/_domain/constants/createTeam.ts
@@ -1,6 +1,3 @@
-/* TODO 무조건적인 true 대신에 로그인한 유저의 정보를 바탕으로 받아오는 로직 작성해야 함*/
-export const ENABLE_NO_TEAM_STATE_PREVIEW = true;
-
 export const CREATE_TEAM_MESSAGES = {
   success: '팀이 생성되었습니다.',
   emptyTeamNameError: '팀 이름을 입력해주세요.',
