+++ src/app/addteam/layout.tsx
@@ -1,6 +1,4 @@
 import { MobileHeader, Sidebar } from '@/components/sidebar';
-import clsx from 'clsx';
-import commonStyles from './_styles/common.module.css';
 import styles from './page.module.css';
 
 export default function AddTeamLayout({ children }: { children: React.ReactNode }) {
@@ -10,7 +8,7 @@ export default function AddTeamLayout({ children }: { children: React.ReactNode 
       <div className={styles.mobileOnlyHeader}>
         <MobileHeader />
       </div>
-      <section className={clsx(commonStyles.flexCenter, styles.mainContents)}>{children}</section>
+      <section className={styles.mainContents}>{children}</section>
     </main>
   );
 }
